+++ apis/api-area001-v003-mal/src/main/java/org/ccsds/moims/mo/mal/helpertools/helpers/HelperMisc.java
@@ -124,6 +124,16 @@ public class HelperMisc {
      * @throws IOException if the file consumer properties file does no exist
      */
     public static void loadConsumerProperties() throws MalformedURLException, IOException {
+        // Were they loaded already? This flag is independent from the provider's
+        // "PropertiesLoadedFlag" so that an application acting as both a consumer
+        // and a provider still loads both property files.
+        String propAreLoaded = System.getProperty("ConsumerPropertiesLoadedFlag");
+        if (propAreLoaded != null) {
+            if (propAreLoaded.equals("true")) {
+                return;
+            }
+        }
+
         final Properties sysProps = System.getProperties();
         final File file = new File(System.getProperty("consumer.properties", CONSUMER_PROPERTIES_FILE));
 
@@ -134,6 +144,7 @@ public class HelperMisc {
         }
 
         System.setProperties(sysProps);
+        System.setProperty("ConsumerPropertiesLoadedFlag", "true");
     }
 
     /**
@@ -218,7 +229,6 @@ public class HelperMisc {
      * SHARED_BROKER_PROPERTIES file will be read
      */
     public static void loadPropertiesFile(Boolean useSharedBroker) {
-
         // Were they loaded already?
         String propAreLoaded = System.getProperty("PropertiesLoadedFlag");
         if (propAreLoaded != null) {
@@ -296,11 +306,9 @@ public class HelperMisc {
 
             System.setProperties(sysProps);
             System.setProperty("PropertiesLoadedFlag", "true");
-
         } catch (MalformedURLException ex) {
             Logger.getLogger(HelperMisc.class.getName()).log(Level.SEVERE, null, ex);
         }
-
     }
 
     /**